requirements.txt
Flask==0.11.1
Flask-HTTPAuth==3.2.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.11.11
appengine==1.8.0.2
beautifulsoup4==4.5.1
click==6.6
google-cloud-core==0.20.0
google-cloud-language==0.20.0
google-cloud-natural-language==0.20.0
googleapis-common-protos==1.3.5
httplib2==0.9.2
itsdangerous==0.24
oauth2client==3.0.0
protobuf==3.1.0.post1
pyasn1==0.1.9
pyasn1-modules==0.0.8
requests==2.11.1
requests-toolbelt==0.7.0
rsa==3.4.2
six==1.10.0
wsgiref==0.1.2
